refactor: useQurey - getContents, deleteContent 적용 - useQurey사용으로 getContents 함수를 deleteButton에 프롭으로 넘기지 않고 queryKey를 통해 호출 했습니다. + 기존 useQurey 삭제,

+++ apps/admin/components/lib/antd/ContentButtons/index.tsx
@@ -9,18 +9,38 @@ import { deleteContentAPI } from '@/services/contents';
 import { initialMbtiTestData, isEditContentState, mbtiTestDataState } from '@/states/contentUpdateState';
 import { messageState } from '@/states/messageState';
 import { MessageState } from '@/types/util';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-export const DeleteButton = ({ testId, handleDeleteBtn }: { testId: string; handleDeleteBtn: () => void }) => {
+export const DeleteButton = ({ testId, handleDeleteBtn }: { testId: string; handleDeleteBtn?: () => void }) => {
   const setMessageState = useSetRecoilState<MessageState>(messageState);
+  const queryClient = useQueryClient();
 
-  const onClickDeleteBtn = async () => {
-    await deleteContentAPI(testId);
-    setMessageState({
-      isOn: true,
-      type: 'success',
-      content: '삭제 완료',
-    });
-    handleDeleteBtn();
+  const { mutate: deleteContent } = useMutation({
+    mutationFn: (testId: string) =>
+      deleteContentAPI(testId).then(() =>
+        setMessageState({
+          isOn: true,
+          type: 'success',
+          content: '삭제 완료',
+        }),
+      ),
+    onSettled: async () => {
+      if (handleDeleteBtn) handleDeleteBtn();
+      return await queryClient.invalidateQueries({ queryKey: ['getContents'] });
+    },
+    onError: (error: Error) => {
+      setMessageState({
+        isOn: true,
+        type: 'error',
+        content: `Error: ${error.message}`,
+      });
+    },
+  });
+
+  const onClickDeleteBtn = () => {
+    if (testId) {
+      deleteContent(testId);
+    }
   };
 
   return (
